+++ test/bigtest/tests/users-show-all-test.js
@@ -91,6 +91,19 @@ describe('Users', () => {
       });
     });
 
+    describe('search by barcode index', function () {
+      beforeEach(async function () {
+        const userItem = users[0];
+        await usersInteractor.chooseSearchOption('Barcode');
+        await usersInteractor.searchField.fill(userItem.barcode);
+        await usersInteractor.searchButton.click();
+      });
+
+      it('should find user with given barcode', () => {
+        expect(usersInteractor.instances().length).to.be.equal(1);
+      });
+    });
+
     describe('click clear status filter', () => {
       beforeEach(async function () {
         await usersInteractor.clearStatusFilter();