+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Allow search by last name. Refs UIU-1706.
 * Add user's fees/fines permissions. Refs UIU-1918.
 * Allow cancel fee/fine as error only if any 'actions' has not been applied. Refs UIU-1894.
+* Allow search by barcode. Refs UIU-1708.
 * Add "Users: User loans anonymize" permission. Refs UIU-1535.
 
 ## [5.0.1](https://github.com/folio-org/ui-users/tree/v5.0.1) (2020-10-15)